Add course code in model and remove year
The 'Course' Model has the course code for each different year and
courses. This also adds new room types for rooms reserved for research
scholars and rooms that can be treated as both double or single. The
'Year' model is being removed because it is redundant as the course
already stores the year.